Refactor writing of input files into a separate function
